+++ package.json
@@ -16,6 +16,7 @@
     "react": "^17.0.2",
     "react-calendar-heatmap": "^1.8.1",
     "react-dom": "^17.0.2",
+    "react-firebase-hooks": "^5.0.2",
     "react-modern-calendar-datepicker": "^3.1.6",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
